Refac: Lobby 페이지 리팩토링
1. FriendsList 디렉토리를 Menu 하위로 넣기
2. Menu 이미지 "%"로 변경

resolved #48
