+++ src/index.css
@@ -15,18 +15,18 @@
 }
 
 html {
-  font-size: 10px;
+  font-size: 16px;
 }
 
 @media screen and (max-width: 1920px) {
   html {
-    font-size: 7px;
+    font-size: 10px;
   }
 }
 
 @media screen and (max-height: 600px) {
   html {
-    font-size: 7px;
+    font-size: 10px;
   }
 }
