+++ src/pages/Game/Canvas/Canvas.module.css
