+++ src/pages/Lobby/Menu/Friends/Avatar.module.css
