+++ src/pages/Lobby/Menu/Friends/Avatar.tsx
