+++ src/pages/Lobby/Menu/Friends/Friends.module.css
