+++ src/pages/Lobby/Menu/Friends/Friends_backup.txt
